refactor: split frontend lecture studio ui sections
